refactor(pwf): move pwf serializer into lookup_tables
